requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==23.1.0
ansi2html==1.8.0
asn1crypto==0.24.0
asttokens==2.2.1
attrs==17.4.0
autoflake==2.2.0
Automat==0.6.0
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.4
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.12
classify-imports==4.2.0
click==8.1.7
colorama==0.3.7
comm==0.1.4
ConfigArgParse==1.7
configobj==5.0.6
constantly==15.1.0
cycler==0.11.0
dash==2.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
decorator==4.4.2
distro-info==1.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.2.5
fvcore==0.1.5.post20221221
gin-config==0.5.0
google-auth==2.20.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==16.10.1
iopath==0.1.10
ipython==8.12.2
ipywidgets==8.1.0
isort==5.12.0
itsdangerous==2.1.2
jaxtyping==0.2.19
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.3.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-service==2.4.0
nbformat==5.7.0
nest-asyncio==1.5.7
netifaces==0.10.4
networkx==2.5.1
ninja==1.11.1
numpy==1.24.3
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.7.0.68
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.16.1
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pyflakes==3.1.0
Pygments==2.16.1
pyhocon==0.3.59
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyquaternion==0.9.9
pyserial==3.4
python-dateutil==2.8.2
python-typing-update==0.6.0
pytz==2023.3
pyupgrade==3.10.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==6.0
raymarching==0.1.2
reorder-python-imports==3.10.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
retrying==1.3.4
rich==13.5.2
rsa==4.9
scikit-image==0.17.2
scikit-learn==1.3.0
scipy==1.5.4
SecretStorage==2.3.1
service-identity==16.0.0
setuptools==59.5.0
six==1.11.0
stack-data==0.6.2
style==1.1.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.2.0
tifffile==2020.9.3
tokenize-rt==5.2.0
tomli==2.0.1
tqdm==4.64.1
traitlets==5.9.0
trimesh==3.22.1
typeguard==4.1.0
typing_extensions==4.7.0
tzdata==2023.3
update==0.0.1
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
yacs==0.1.8
zipp==3.16.2
